requirements.txt
bs4==0.0.1
PyJWT==1.7.1
Django==2.1.5
boto3==1.12.40
urllib3==1.25.3
requests==2.23.0
drf-yasg==1.17.0
selenium==3.141.0
django-silk==3.0.2
mysqlclient==1.3.14
google-auth==1.14.1
django-filter==2.1.0
python-dateutil==2.8.0
optional-django==0.1.0
psycopg2-binary==2.8.5
django-storages==1.9.1
django-rest-knox==4.1.0
google-api-core==1.17.0
google-api-helper==0.3.1
django-rest-swagger==2.2.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-cors-headers==3.2.1
djangorestframework==3.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
django-cool-pagination==0.3.0
mysql-connector-python==8.0.20
google-api-python-client==1.8.2
googleapis-common-protos==1.51.0 
djangorestframework-simplejwt==3.3
